+++ systemctl_mqtt/__init__.py
@@ -15,6 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+import abc
 import argparse
 import datetime
 import functools
@@ -50,6 +51,7 @@ class _State:
         mqtt_topic_prefix: str,
         homeassistant_discovery_prefix: str,
         homeassistant_node_id: str,
+        poweroff_delay: datetime.timedelta,
     ) -> None:
         self._mqtt_topic_prefix = mqtt_topic_prefix
         self._homeassistant_discovery_prefix = homeassistant_discovery_prefix
@@ -59,6 +61,7 @@ class _State:
         )  # type: dbus.proxies.Interface
         self._shutdown_lock = None  # type: typing.Optional[dbus.types.UnixFd]
         self._shutdown_lock_mutex = threading.Lock()
+        self.poweroff_delay = poweroff_delay
 
     @property
     def mqtt_topic_prefix(self) -> str:
@@ -192,13 +195,10 @@ class _State:
         )
 
 
-class _MQTTAction:
-
-    # pylint: disable=too-few-public-methods
-
-    def __init__(self, name: str, action: typing.Callable) -> None:
-        self.name = name
-        self.action = action
+class _MQTTAction(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def trigger(self, state: _State) -> None:
+        pass  # pragma: no cover
 
     def mqtt_message_callback(
         self,
@@ -213,21 +213,23 @@ class _MQTTAction:
         if message.retain:
             _LOGGER.info("ignoring retained message")
             return
-        _LOGGER.debug("executing action %s (%r)", self.name, self.action)
-        self.action()
-        _LOGGER.debug("completed action %s (%r)", self.name, self.action)
+        _LOGGER.debug("executing action %s", self)
+        self.trigger(state=state)
+        _LOGGER.debug("completed action %s", self)
 
 
-_MQTT_TOPIC_SUFFIX_ACTION_MAPPING = {
-    "poweroff": _MQTTAction(
-        name="poweroff",
-        action=functools.partial(
-            # pylint: disable=protected-access
-            systemctl_mqtt._dbus.schedule_shutdown,
-            action="poweroff",
-        ),
-    ),
-}
+class _MQTTActionSchedulePoweroff(_MQTTAction):
+    def trigger(self, state: _State) -> None:
+        # pylint: disable=protected-access
+        systemctl_mqtt._dbus.schedule_shutdown(
+            action="poweroff", delay=state.poweroff_delay
+        )
+
+    def __str__(self) -> str:
+        return type(self).__name__
+
+
+_MQTT_TOPIC_SUFFIX_ACTION_MAPPING = {"poweroff": _MQTTActionSchedulePoweroff()}
 
 
 def _mqtt_on_connect(
@@ -254,7 +256,7 @@ def _mqtt_on_connect(
             sub=topic, callback=action.mqtt_message_callback
         )
         _LOGGER.debug(
-            "registered MQTT callback for topic %s triggering %r", topic, action.action,
+            "registered MQTT callback for topic %s triggering %s", topic, action
         )
 
 
@@ -266,6 +268,7 @@ def _run(
     mqtt_topic_prefix: str,
     homeassistant_discovery_prefix: str,
     homeassistant_node_id: str,
+    poweroff_delay: datetime.timedelta,
     mqtt_disable_tls: bool = False,
 ) -> None:
     # pylint: disable=too-many-arguments
@@ -277,6 +280,7 @@ def _run(
             mqtt_topic_prefix=mqtt_topic_prefix,
             homeassistant_discovery_prefix=homeassistant_discovery_prefix,
             homeassistant_node_id=homeassistant_node_id,
+            poweroff_delay=poweroff_delay,
         )
     )
     mqtt_client.on_connect = _mqtt_on_connect
@@ -354,6 +358,9 @@ def _main() -> None:
         default=systemctl_mqtt._homeassistant.get_default_node_id(),
         help=" ",
     )
+    argparser.add_argument(
+        "--poweroff-delay-seconds", type=float, default=4.0, help=" ",
+    )
     args = argparser.parse_args()
     if args.mqtt_port:
         mqtt_port = args.mqtt_port
@@ -389,4 +396,5 @@ def _main() -> None:
         mqtt_topic_prefix=args.mqtt_topic_prefix,
         homeassistant_discovery_prefix=args.homeassistant_discovery_prefix,
         homeassistant_node_id=args.homeassistant_node_id,
+        poweroff_delay=datetime.timedelta(seconds=args.poweroff_delay_seconds),
     )