+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- command line option `--poweroff-delay-seconds` (default: 4 seconds)
+
 ### Changed
 - docker build stage: revert user after applying `chown` workaround for inter-stage copy